feat(precond): light_result + float-FFT preconditioner for lean large-Ns solves
- solve_contact/solve_contact_impl gain light=: skip displacement/gap result
  arrays (2 of 3 double N-vectors). nested passes it (coarse levels always
  light); exposed as solve_nested(light_result=).
- FourierPreconditioner symbol stored float; apply/apply_single share a
  scalar-templated float/double FFT. apply_single now runs a real float FFT
  (halved transients) -> single precision keeps the preconditioner ON, which it
  needs to converge.
- nested frees each level's double gap after use.
Measured: single_precision+light_result ~halves the solve (Ns=4096 1.37->0.74
GiB). Validated at Ns=512: converges, matches double to dArea 1.1e-5.

+++ include/fourier_precond.hpp
@@ -27,7 +27,7 @@ public:
 
 private:
     int Ns_;
-    Eigen::MatrixXd w_; // Ns×Ns wavenumber symbol |k|, w_(0,0)=0
+    Eigen::MatrixXf w_; // Ns×Ns wavenumber symbol |k| (float), w_(0,0)=0
 };
 
 } // namespace hmc
